pyproject.toml
[project]
name = "pyne"
version = "0.1.0"
description = "AI-oli - Jaffle Shop Assistant chatbot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "dash>=4.0.0",
    "duckdb>=1.5.0",
    "huggingface-hub>=1.9.0",
    "mcp>=1.26.0",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
